Port Basis tests to use doctest.
